requirements.txt
flask
gunicorn
google-generativeai
flask
gunicorn
